+++ R/exfm22.R
@@ -1,0 +1,20 @@
+#' @title 
+#' Revenue data if an eucalyptus forest
+#' @description 
+#' In this data, each observation is the year's revenue.
+#' @usage 
+#' data(exfm22)
+#' 
+#' @format
+#' A data frame with 8 observations and 3 variables:
+#' \describe{
+#'   \item{year}{Revenue year}
+#'   \item{cost}{cost values for that year, in dollars}
+#'   \item{revenue}{revenue values for that year, in dollars}
+#' }
+#'
+#' @name exfm22
+#' @docType data
+#' @author Sollano Rabelo Braga \email{sollanorb@@gmail.com}
+#' @keywords data
+NULL